+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Handle refresh token flow for HttpOnly session cookies [#3759](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3759)
 - Handle missing refresh token for HttpOnly session cookies [#3771](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3771)
 - Add HttpOnly session cookies for SLAS public client [#3774](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3774)
+- Ensure client-side refresh token requests are deduped [#3786](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3786/)
 - Add USID support to StorefrontPreview for HttpOnly cookies [#3785](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3785)
 
 ## v5.2.0-dev (Mar 20, 2026)